requirements.txt
aiogram==3.20.0.post0
telethon==1.40.0
sqlalchemy==2.0.41
python-dotenv==1.1.0
ollama==0.5.1
requests==2.32.3